+++ ChangeLog.md
@@ -1,10 +1,29 @@
 # Change Log for CodeIgniter Simple and Secure Twig
 
-## v0.3.0 (Not Released)
+## v1.0.0 (2021/06/01)
+
+* Require Twig 3.0 or newer;
+* Refactor to Use Autoloader/PSR-4
+* Refactor tests to use Autoloader
+## v0.4.0 (2019/03/22)
+
+### Changed
+
+* Require Twig 1.38.0 or later.
+
+### Added
+
+* No direct script access allowed.
+
+## v0.3.0 (2018/08/14)
 
 ### Added
 
 * Add functionality to override any Twig Environment options.
+
+### Changed
+
+* Require Twig 1.22.0 or later.
 
 ## v0.2.1 (2016/06/17)